+++ node/src/types/kernel/api/notebook/setNotebookConf/response.d.ts
@@ -48,6 +48,10 @@ export interface IData {
      */
     dailyNoteTemplatePath: string;
     /**
+     * New document save notebook
+     */
+    docCreateSaveBox: string;
+    /**
      * New document save location
      */
     docCreateSavePath: string;
@@ -60,7 +64,11 @@ export interface IData {
      */
     name: string;
     /**
-     * Ref create doc save location
+     * The notebook that was stored when a new document was created using block references
+     */
+    refCreateSaveBox: string;
+    /**
+     * The document path that was stored when a new document was created using block references
      */
     refCreateSavePath: string;
     /**